requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
attrs==21.4.0
autopep8==1.6.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
commonmark==0.9.1
Deprecated==1.2.13
filelock==3.6.0
google-api-core==2.7.1
google-api-python-client==2.41.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
httplib2==0.20.4
huggingface-hub==0.4.0
idna==3.3
iniconfig==1.1.1
joblib==1.1.0
keybert==0.5.0
nltk==3.7
numpy==1.22.3
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
protobuf==3.19.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
Pygments==2.11.2
pyparsing==3.0.7
pytest==7.1.1
python-dotenv==0.19.2
PyYAML==6.0
redis==4.1.4
regex==2022.3.15
requests==2.27.1
rich==12.0.0
rq==1.10.1
rsa==4.8
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.8.0
sentence-transformers==2.2.0
sentencepiece==0.1.96
six==1.16.0
slack-bolt==1.13.0
slack-sdk==3.15.2
threadpoolctl==3.1.0
tokenizers==0.11.6
toml==0.10.2
tomli==2.0.1
torch==1.10.0+cpu
torchaudio==0.10.0+cpu
torchvision==0.11.0+cpu
tqdm==4.63.0
transformers==4.17.0
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.0
testresources==2.0.1